refactor(cli): split remote/index.ts into 9 focused modules
- model-selection.ts: setModelFromWeb handler, model registry lookup
- session-name-sync.ts: startSessionNameSync, stopSessionNameSync, polling state
- trigger-cancellation.ts: cancellation retry loop and batch state
- delink-management.ts: emitDelinkChildren, emitDelinkOwnParent, retry timers
- followup-grace.ts: startFollowUpGrace, clearFollowUpGrace, fireSessionComplete
- relay-context-factory.ts: RelayContext object, forwardEvent, sendToWeb, etc.
- lifecycle-handlers.ts: all pi event listeners
- connection-handlers-factory.ts: ConnectionHandlers callbacks, doConnect/doDisconnect
- index.ts: thin orchestrator (~80 lines), module-level state, public accessors

Existing sub-modules (chunked-delivery.ts, registration-gate.ts, connection.ts)
are untouched. All existing imports continue to work via re-exports.

All 1418 CLI tests pass.
